Place bandaide firmly over failing doctest
Weirdest bug I've ever seen: On random occasions, the doctest would fail
with a ZeroDivisionError, stating that the AffineCipher could not find
the modular inverse of the given alpha value. This is weird because the
doctest should always get an alpha value of 9 (there's nothing
probabilistic in the naive attack). Instead, it would sometimes find the
solution pair as (20,22).

WTF